4_Test_Automation/homework_pytest/requirements.txt
pytest==7.2.2
pymssql==2.2.7
python-dotenv==1.0.0
pytest-html==3.2.0
